[FEAT] 정보 수정제안 API 연결

+++ NanaLand/NanaLand/ReportInfo/ReportInfoDTO.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct ReportInfoRequest: Codable {
-	let postId: Int
+	let postId: Int64
 	let fixType: String
 	let category: String
 	let content: String
